requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6

# Dashboard
streamlit==1.28.2
plotly==5.18.0
altair==5.1.2

# Security
cryptography==41.0.7
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pillow==10.1.0
requests==2.31.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
